chore: have addresses package simple build to esm and rexports addresses from dma-deployment
