--- packages/addresses/tsconfig.base.json
@@ -1,27 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2020",
-    "moduleResolution": "node",
-    "strict": true,
-    "baseUrl": ".",
-    "types": ["node"],
-    "forceConsistentCasingInFileNames": true,
-    "allowSyntheticDefaultImports": true,
-    "esModuleInterop": true,
-    "allowJs": true,
-    "declaration": true,
-    "declarationMap": true,
-    "sourceMap": true,
-    "resolveJsonModule": true,
-    "noImplicitAny": false,
-    "typeRoots": ["../../node_modules/@types"],
-    "paths": {
-      /* Aliases used internally by referenced packages */
-      "@dma-deployments/utils/*": ["../dma-deployments/utils/*"],
-      "@dma-deployments/types/*": ["../dma-deployments/types/*"],
-      "@dma-deployments/constants": ["../dma-deployments/constants/index.ts"]
-    }
-  },
-  "include": ["./src/**/*"],
-  "exclude": ["node_modules", "../dma-deployments"],
-}
