--- packages/addresses/tsconfig.esm.json
@@ -1,7 +1,0 @@
-{
-  "extends": "./tsconfig.base.json",
-  "compilerOptions": {
-    "module": "ES2020",
-    "outDir": "lib/esm"
-  },
-}
