+++ packages/addresses/tsconfig.json
@@ -1,0 +1,48 @@
+{
+  "compilerOptions": {
+    "target": "ES2020",
+    "moduleResolution": "node",
+    "strict": true,
+    "incremental": false,
+    "baseUrl": ".",
+    "rootDir": "./src",
+    "outDir": "./lib",
+    "types": [
+      "node"
+    ],
+    "forceConsistentCasingInFileNames": true,
+    "allowSyntheticDefaultImports": true,
+    "esModuleInterop": true,
+    "allowJs": true,
+    "declaration": true,
+    "declarationMap": true,
+    "emitDeclarationOnly": true,
+    "sourceMap": true,
+    "resolveJsonModule": true,
+    "noImplicitAny": false,
+    "typeRoots": [
+      "../../node_modules/@types"
+    ],
+    "paths": {
+      /* Aliases used internally by referenced packages */
+      "@dma-deployments/configs/*": [
+        "../dma-deployments/configs/*"
+      ],
+      "@dma-deployments/utils/*": [
+        "../dma-deployments/utils/*"
+      ],
+      "@dma-deployments/types/*": [
+        "../dma-deployments/types/*"
+      ],
+      "@dma-deployments/constants": [
+        "../dma-deployments/constants/index.ts"
+      ]
+    }
+  },
+  "include": [
+    "./src/index.ts"
+  ],
+  "exclude": [
+    "node_modules"
+  ]
+}
