+++ packages/dma-common/types/address.ts
@@ -1,1 +1,1 @@
-export { Address } from '@oasisdex/dma-deployments/types/address'
+export type { Address } from '@oasisdex/dma-deployments/types/address'
