+++ packages/dma-contracts/test/unit/swap/u-swap.test.ts
@@ -1,3 +1,4 @@
+import { amountToWei } from '@dma-common/utils/common'
 import { createDeploy } from '@dma-common/utils/deploy'
 import init from '@dma-common/utils/init'
 import { swap, uniswapV3Swap, unoswap } from '@dma-contracts/test/fixtures'
@@ -6,7 +7,7 @@ import { ADDRESSES } from '@oasisdex/addresses'
 import { CONTRACT_NAMES, ONE } from '@oasisdex/dma-common/constants'
 import { asPercentageValue, expect, FEE } from '@oasisdex/dma-common/test-utils'
 import { RuntimeConfig } from '@oasisdex/dma-common/types/common'
-import { amountToWei, balanceOf } from '@oasisdex/dma-common/utils/common'
+import { balanceOf } from '@oasisdex/dma-common/utils/balances'
 import { calculateFee, swapOneInchTokens } from '@oasisdex/dma-common/utils/swap'
 import { Network } from '@oasisdex/dma-deployments/types/network'
 import { ServiceRegistry } from '@oasisdex/dma-deployments/utils/wrappers'
