+++ packages/dma-deployments/configs/index.ts
@@ -1,0 +1,4 @@
+export { goerliConfig } from './goerli.conf'
+export { localExtendConfig } from './local-extend.conf'
+export { mainnetConfig } from './mainnet.conf'
+export { optimismConfig } from './optimism.conf'
