+++ packages/dma-deployments/package.json
@@ -2,16 +2,10 @@
   "name": "@oasisdex/dma-deployments",
   "version": "1.0.0",
   "license": "Apache-2.0",
+  "main": "./addresses/index.ts",
+  "types": "./addresses/index.ts",
   "scripts": {
     "lint": "yarn tsc -p ./tsconfig.json --noEmit"
-  },
-  "exports": {
-    "./configs/*": "./configs/*.ts",
-    "./constants": "./constants/index.ts",
-    "./operation-definitions": "./operation-definitions/index.ts",
-    "./deployment/*": "./deployment/*.ts",
-    "./types/*": "./types/*.ts",
-    "./utils/*": "./utils/*/index.ts"
   },
   "dependencies": {
     "@oasisdex/abis": "*",
